+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Added
 * Support for Sentinel-2 L2A COG processing via the AWS Element84 API.
+* Option to force the regeneration of Sentinel-1 static layers.
 
 ### Changed
 * `process.get_s2_metadata` now dynamically routes to either the Google Cloud Storage L1C workflow or the new AWS L2A workflow based on the scene name format.